move things a little bit around to perhaps make it slightly more logical
